--- safe/README.md
@@ -1,11 +1,0 @@
-# sketch-youtube
-
-[Work in progress]
-
-## Experimental!
-
-This app was created during [Matt](https://interconnected.org)'s summer 2023 residency. The purpose is to experiment with multiplayer interactions, and simultaneously see what PartyKit can do. It's called a sketch because it's lightweight and quick, and because we learn something in making it.
-
-## To do
-
-- [ ] Needs tap-to-unmute on first load
